WIP namespacing proposal
