+++ sdk-blocking/src/main/java/dev/restate/sdk/blocking/RestateBlockingService.java
@@ -12,4 +12,8 @@ public interface RestateBlockingService extends BindableBlockingService {
   default RestateContext restateContext() {
     return new RestateContextImpl(Syscalls.current());
   }
+
+  default RestateInstanceStorage instanceStorage() {
+    return restateContext().storage();
+  }
 }
